==> r-data-cleaning-and-visualization/1. RMD files/1. Education data cleaning.Rmd <==
---
title: "Education data cleaning"
author: "Stephanie Veazie"
date: "2022-12-02"
output:
  word_document: default
  html_document: default
---

## Download data 

1. Go to https://data.census.gov/cedsci/

2. Select "advanced search." 

3. Using the filter pane on the left-hand side of the page, under "Geography", select "State" then "All states with in the US, Puerto Rico, and the Island Areas." 

4. Under "Topics," select "Education" then "Educational attainment."

5. Under "Topics" select "Race and ethnicity" then "Race and Ethnicity" again.

6. Under "Surveys" select "American Community Survey", then "5-Year Estimates" then "Detailed Tables" 

7. Click blue "Search" button in the bottom right-hand corner.

8. Select "C15002H" for White Alone, Not Hispanic or Latino. 

9. Under drop-down menu, select "2019: ACS 5-year estimates subject tables"

10. Select "transpose."

11. Download as a csv file. 

12. Repeat steps 1-7 but select "C15002B" for Black Alone. 

## Load libraries

```{r}
# Load libraries
library(tidyverse)
library(dplyr)
```

## Import & clean data on White NH pops

```{r}
# Import data on white nh pop
edu_data_white<-read_csv("../2. Raw data files/acs_education_by_state_white_nh_2015-2019.csv")
```

```{r}
# Simplify column names & add race label
colnames(edu_data_white)<-c(
  "variable",
  "total_white",
  "total_male_white",
  "male_less_than_high_school_white",
  "male_high_school_white",
  "male_some_college_white",
  "male_college_or_higher_white", 
  "total_female_white",
  "female_less_than_high_school_white",
  "female_high_school_white",
  "female_some_college_white",
  "female_college_or_higher_white")

# Confirm colnames were changed
colnames(edu_data_white)
```

```{r}
# Create vectors for specific states. Specify r (# of rows/state.) 
# NOTE - Puerto Rico comes at the end of the list for all ACS datasets.

r<- 3

alabama_vector<-rep("Alabama", times= r)
alaska_vector<-rep("Alaska", times=r)
arizona_vector<-rep("Arizona", times=r)
arkansas_vector<-rep("Arkansas", times=r)
california_vector<-rep("California", times=r)
colorado_vector<-rep("Colorado", times=r)
connecticut_vector<-rep("Connecticut", times=r)
delaware_vector<-rep("Delaware", times=r)
dc_vector<-rep("District of Columbia", times=r)
florida_vector<-rep("Florida", times=r)
georgia_vector<-rep("Georgia", times=r)
hawaii_vector<-rep("Hawaii", times=r)
idaho_vector<-rep("Idaho", times=r)
illinois_vector<-rep("Illinois", times=r)
indiana_vector<-rep("Indiana", times=r)
iowa_vector<-rep("Iowa", times=r)
kansas_vector<-rep("Kansas", times=r)
kentucky_vector<-rep("Kentucky", times=r)
louisiana_vector<-rep("Louisiana", times=r)
maine_vector<-rep("Maine", times=r)
maryland_vector<-rep("Maryland", times=r)
massachusetts_vector<-rep("Massachusetts", times=r)
michigan_vector<-rep("Michigan", times=r)
minnesota_vector<-rep("Minnesota", times=r)
mississippi_vector<-rep("Mississippi", times=r)
missouri_vector<-rep("Missouri", times=r)
montana_vector<-rep("Montana", times=r)
nebraska_vector<-rep("Nebraska", times=r)
nevada_vector<-rep("Nevada", times=r)
newhampshire_vector<-rep("New Hampshire", times=r)
newjersey_vector<-rep("New Jersey", times=r)
newmexico_vector<-rep("New Mexico", times=r)
newyork_vector<-rep("New York", times=r)
northcarolina_vector<-rep("North Carolina", times=r)
northdakota_vector<-rep("North Dakota", times=r)
ohio_vector<-rep("Ohio", times=r)
oklahoma_vector<-rep("Oklahoma", times=r)
oregon_vector<-rep("Oregon", times=r)
pennsylvania_vector<-rep("Pennsylvania", times=r)
rhodeisland_vector<-rep("Rhode Island", times=r)
southcarolina_vector<-rep("South Carolina", times=r)
southdakota_vector<-rep("South Dakota", times=r)
tennessee_vector<-rep("Tennessee", times=r)
texas_vector<-rep("Texas", times=r)
utah_vector<-rep("Utah", times=r)
vermont_vector<-rep("Vermont", times=r)
virginia_vector<-rep("Virginia", times=r)
washington_vector<-rep("Washington", times=r)
westvirginia_vector<-rep("West Virginia", times=r)
wisconsin_vector<-rep("Wisconsin", times=r)
wyoming_vector<-rep("Wyoming", times=r)
puertorico_vector<-rep("Puerto Rico", times=r)
```

```{r}
# Create a single state vector made up of all the individual state vectors. 
state_vector<-c(alabama_vector, 
                alaska_vector, 
                arizona_vector,
                arkansas_vector, 
                california_vector,
                colorado_vector,
                connecticut_vector,
                delaware_vector,
                dc_vector,
                florida_vector, 
                georgia_vector,
                hawaii_vector,
                idaho_vector, 
                illinois_vector,
                indiana_vector, 
                iowa_vector, 
                kansas_vector,
                kentucky_vector,
                louisiana_vector,
                maine_vector,
                maryland_vector,
                massachusetts_vector, 
                michigan_vector,
                minnesota_vector,
                mississippi_vector,
                missouri_vector,
                montana_vector,
                nebraska_vector,
                nevada_vector,
                newhampshire_vector,
                newjersey_vector,
                newmexico_vector,
                newyork_vector,
                northcarolina_vector,
                northdakota_vector,
                ohio_vector,
                oklahoma_vector,
                oregon_vector,
                pennsylvania_vector,
                rhodeisland_vector,
                southcarolina_vector,
                southdakota_vector,
                tennessee_vector,
                texas_vector,
                utah_vector,
                vermont_vector,
                virginia_vector,
                washington_vector,
                westvirginia_vector,
                wisconsin_vector,
                wyoming_vector,
                puertorico_vector)

# Confirm all states, DC & PR were included. This should = 52.
length(state_vector) / r 
```

```{r}
# Add state vector to selected data set
edu_data_white<- cbind(state_vector, edu_data_white)

# Confirm change
colnames(edu_data_white)
```
```{r}
# Create individual vectors with the row numbers of interest (row 2-estimate). Each vector corresponds to the rows for a specific state. Create empty, master vector.

initial_vector = c(2)
master_vector = c()

# Loop through each of the 52 states, and for each one use the iteration number of the loop as part of the equation. Then, append the resulting vector to the master vector. 
for (i in 1:52){
  vector = initial_vector+r*(i-1)
  master_vector = append(master_vector, vector)}

# Rename the vector
row_numbers_education <- master_vector

# Confirm row numbers make sense
row_numbers_education
```

```{r}
# Subset dataset to 1 observation of interest for each state, based on loop above. 
tidy_education<-edu_data_white[row_numbers_education, ]

# Confirm dataset is 52 rows long
tidy_education
```

```{r}
# Remove column that just has observations that say "estimate"
tidy_education<- tidy_education[, -2]

# Confirm change was made
tidy_education
```

```{r}
# Remove ',' and '%' from data in order to create numerical data
for (i in 2:12){
  tidy_education[,i] = gsub("%", "", tidy_education[,i])
  tidy_education[,i] = gsub(",", "", tidy_education[,i])
  tidy_education[,i] = as.numeric(tidy_education[,i])}

# Confirm change 
tidy_education
```

```{r}
# Limit to total # and # college educated columns
tidy_education<- tidy_education[,c(
  "state_vector",
  "total_white", 
  "male_college_or_higher_white",
  "female_college_or_higher_white"
)]

# Confirm change
tidy_education
```

```{r}
# Calculate % of pop w/out college degree
tidy_education_white<- mutate(tidy_education, 
                        percent_no_college_white = 
                          1- 
                             ((male_college_or_higher_white +
                                 female_college_or_higher_white) / 
                                total_white)
                        )

# Confirm changes was made
tidy_education_white
```

# Import and clean data on Black pops

```{r}
# Import data on black pop
edu_data_black<-read_csv("../2. Raw data files/acs_education_by_state_black_2015-2019.csv")

colnames(edu_data_black)
```

```{r}
# Simplify column names & add race label
colnames(edu_data_black)<-c(
  "variable",
  "total_black",
  "total_male_black",
  "male_less_than_high_school_black",
  "male_high_school_black",
  "male_some_college_black",
  "male_college_or_higher_black", 
  "total_female_black",
  "female_less_than_high_school_black",
  "female_high_school_black",
  "female_some_college_black",
  "female_college_or_higher_black")

# Confirm colnames were changed
edu_data_black
```

```{r}
# Create vectors for specific states. Specify r (# of rows/state.) 
# NOTE - Puerto Rico comes at the end of the list for all ACS datasets.

r<- 3

alabama_vector<-rep("Alabama", times= r)
alaska_vector<-rep("Alaska", times=r)
arizona_vector<-rep("Arizona", times=r)
arkansas_vector<-rep("Arkansas", times=r)
california_vector<-rep("California", times=r)
colorado_vector<-rep("Colorado", times=r)
connecticut_vector<-rep("Connecticut", times=r)
delaware_vector<-rep("Delaware", times=r)
dc_vector<-rep("District of Columbia", times=r)
florida_vector<-rep("Florida", times=r)
georgia_vector<-rep("Georgia", times=r)
hawaii_vector<-rep("Hawaii", times=r)
idaho_vector<-rep("Idaho", times=r)
illinois_vector<-rep("Illinois", times=r)
indiana_vector<-rep("Indiana", times=r)
iowa_vector<-rep("Iowa", times=r)
kansas_vector<-rep("Kansas", times=r)
kentucky_vector<-rep("Kentucky", times=r)
louisiana_vector<-rep("Louisiana", times=r)
maine_vector<-rep("Maine", times=r)
maryland_vector<-rep("Maryland", times=r)
massachusetts_vector<-rep("Massachusetts", times=r)
michigan_vector<-rep("Michigan", times=r)
minnesota_vector<-rep("Minnesota", times=r)
mississippi_vector<-rep("Mississippi", times=r)
missouri_vector<-rep("Missouri", times=r)
montana_vector<-rep("Montana", times=r)
nebraska_vector<-rep("Nebraska", times=r)
nevada_vector<-rep("Nevada", times=r)
newhampshire_vector<-rep("New Hampshire", times=r)
newjersey_vector<-rep("New Jersey", times=r)
newmexico_vector<-rep("New Mexico", times=r)
newyork_vector<-rep("New York", times=r)
northcarolina_vector<-rep("North Carolina", times=r)
northdakota_vector<-rep("North Dakota", times=r)
ohio_vector<-rep("Ohio", times=r)
oklahoma_vector<-rep("Oklahoma", times=r)
oregon_vector<-rep("Oregon", times=r)
pennsylvania_vector<-rep("Pennsylvania", times=r)
rhodeisland_vector<-rep("Rhode Island", times=r)
southcarolina_vector<-rep("South Carolina", times=r)
southdakota_vector<-rep("South Dakota", times=r)
tennessee_vector<-rep("Tennessee", times=r)
texas_vector<-rep("Texas", times=r)
utah_vector<-rep("Utah", times=r)
vermont_vector<-rep("Vermont", times=r)
virginia_vector<-rep("Virginia", times=r)
washington_vector<-rep("Washington", times=r)
westvirginia_vector<-rep("West Virginia", times=r)
wisconsin_vector<-rep("Wisconsin", times=r)
wyoming_vector<-rep("Wyoming", times=r)
puertorico_vector<-rep("Puerto Rico", times=r)
```

```{r}
# Create a single state vector made up of all the individual state vectors. 
state_vector<-c(alabama_vector, 
                alaska_vector, 
                arizona_vector,
                arkansas_vector, 
                california_vector,
                colorado_vector,
                connecticut_vector,
                delaware_vector,
                dc_vector,
                florida_vector, 
                georgia_vector,
                hawaii_vector,
                idaho_vector, 
                illinois_vector,
                indiana_vector, 
                iowa_vector, 
                kansas_vector,
                kentucky_vector,
                louisiana_vector,
                maine_vector,
                maryland_vector,
                massachusetts_vector, 
                michigan_vector,
                minnesota_vector,
                mississippi_vector,
                missouri_vector,
                montana_vector,
                nebraska_vector,
                nevada_vector,
                newhampshire_vector,
                newjersey_vector,
                newmexico_vector,
                newyork_vector,
                northcarolina_vector,
                northdakota_vector,
                ohio_vector,
                oklahoma_vector,
                oregon_vector,
                pennsylvania_vector,
                rhodeisland_vector,
                southcarolina_vector,
                southdakota_vector,
                tennessee_vector,
                texas_vector,
                utah_vector,
                vermont_vector,
                virginia_vector,
                washington_vector,
                westvirginia_vector,
                wisconsin_vector,
                wyoming_vector,
                puertorico_vector)

# Confirm all states, DC & PR were included. This should = 52.
length(state_vector) / r 
```

```{r}
# Add state vector to selected data set
edu_data_black<- cbind(state_vector, edu_data_black)

# Confirm change was made
edu_data_black
```

```{r}
# Create individual vectors with the row numbers of interest (row 2-estimate). Each vector corresponds to the rows for a specific state. Create empty, master vector.

initial_vector = c(2)
master_vector = c()

# Loop through each of the 52 states, and for each one use the iteration number of the loop as part of the equation. Then, append the resulting vector to the master vector. 
for (i in 1:52){
  vector = initial_vector+r*(i-1)
  master_vector = append(master_vector, vector)}

# Rename the vector
row_numbers_education <- master_vector
```

```{r}
# Subset to 1 observation of interest for each state, based on loop above. 
tidy_education_black<-edu_data_black[row_numbers_education, ]

# Confirm change was made
tidy_education_black
```

```{r}
# Remove the variable vector
tidy_education_black<- tidy_education_black[, -2]

# Confirm change was made
tidy_education_black
```

```{r}
# Remove ',' and '%' from data in order to create numerical data
for (i in 2:12){
  tidy_education_black[,i] = gsub("%", "", tidy_education_black[,i])
  tidy_education_black[,i] = gsub(",", "", tidy_education_black[,i])
  tidy_education_black[,i] = as.numeric(tidy_education_black[,i])}

# Confirm change was made
tidy_education_black
```

```{r}
# Limit to total # and # college educated columns
tidy_education_black<- tidy_education_black[,c(
  "state_vector",
  "total_black", 
  "male_college_or_higher_black",
  "female_college_or_higher_black"
)]

#Confirm change was made
tidy_education_black
```

```{r}
# Calculate % of pop w/out college degree
tidy_education_black<- mutate(tidy_education_black, 
                        percent_no_college_black = 
                          1 - 
                          ((male_college_or_higher_black +
                            female_college_or_higher_black) / 
                          total_black)
                        )

# Confirm change was made
tidy_education_black
```

```{r}
# Merge datasets
tidy_education_all <- merge(tidy_education_white, tidy_education_black, by.x = "state_vector", by.y = "state_vector")

# Create new column with Black-White ratio of % w/ college degree
tidy_education_all<- mutate(tidy_education_all, 
                            edu_RR = 
                            percent_no_college_black / 
                              percent_no_college_white)

# Confirm changes were made
tidy_education_all
```

```{r}
# Export dataset of all races
write.csv(tidy_education_all,"../3. Cleaned data files/tidy_no_college_25_and_older_3-30-24.csv", row.names = FALSE)
```
